Добавил команду, которая напоминает о чём-либо каждый день Добавил возможность ввода команды в напоминание

+++ apps/service/management/commands/check_notify.py
@@ -2,6 +2,8 @@ from datetime import datetime, timedelta
 
 from django.core.management.base import BaseCommand
 
+from apps.API_VK.VkBotClass import parse_msg
+from apps.API_VK.VkEvent import VkEvent
 from apps.service.models import Notify
 from xoma163site.wsgi import vk_bot
 
@@ -25,4 +27,26 @@ class Command(BaseCommand):
                     vk_bot.send_message(notify.chat.chat_id, message)
                 else:
                     vk_bot.send_message(notify.author.user_id, message)
-                notify.delete()
+
+                # Если отложенная команда
+                if notify.text.startswith('/'):
+                    msg = notify.text[1:]
+                    vk_event = {
+                        'parsed': parse_msg(msg),
+                        'sender': notify.author,
+                    }
+                    if notify.from_chat:
+                        vk_event['chat'] = notify.chat
+                        vk_event['peer_id'] = notify.chat.chat_id
+                    else:
+                        vk_event['chat'] = None
+                        vk_event['peer_id'] = notify.author.user_id
+
+                    vk_event_object = VkEvent(vk_event)
+                    vk_bot.menu(vk_event_object, send=True)
+
+                if notify.repeat:
+                    notify.date = notify.date + timedelta(days=1)
+                    notify.save()
+                else:
+                    notify.delete()
